+++ examples/walletconnect/src/components/BalanceCard.tsx
