+++ examples/walletconnect/src/components/RequestApprovalModal.tsx
