+++ examples/walletconnect/src/components/SessionProposalModal.tsx
