+++ examples/walletconnect/src/hooks/usePolling.ts
