+++ examples/walletconnect/src/hooks/useWalletKit.ts
