+++ examples/walletconnect/src/pages/SettingsPage.tsx
