+++ examples/walletconnect/src/walletkit/auth.ts
