+++ examples/walletconnect/src/walletkit/client.ts
