+++ examples/walletconnect/src/walletkit/handler.ts
