+++ examples/walletconnect/src/walletkit/index.ts
