+++ examples/walletconnect/tsconfig.app.json
