+++ examples/walletconnect/tsconfig.json
