+++ examples/walletconnect/tsconfig.node.json
