+++ examples/walletconnect/vite.config.ts
